+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   their stored properties.
 - Added `split` filter
 - Allow default string filters to be applied to arrays
+- Similar filters are suggested when unknown filter is used
 - Added `indent` filter
 
 ### Bug Fixes